spotify-tracker/requirements.txt
loguru==0.7.3
PyYAML==6.0.2
websockets==12.0
